MAX: Разделил процессоры
